feat(plot_opo): compute the Hausdorff Distance between plots
As a more concrete, mathematical metric of comparing similarity between
the two plots, we use the Hausdorff Distance [1]. Unlike typical
statistical distance metrics, which are more focused on evaluating the
distance between two probability distributions, Hausdorff distance
measures how far two subsets of a metric space (in our case, Euclidean
space) are from each other.

We get a value [0,1] of this metric by normalizing the data points
before processing. We do this by first taking the log10 in order to
match the shape of the log10-based figure, and then subtract by the
minimum values of each dimension, and divide by the maximum values of
each dimension.

[1]: https://en.wikipedia.org/wiki/Hausdorff_distance

See also:

@article{huttenlocher1993comparing,
  title={Comparing images using the Hausdorff distance},
  author={Huttenlocher, Daniel P and Klanderman, Gregory A and Rucklidge, William A},
  journal={IEEE Transactions on Pattern Analysis \& Machine Intelligence},
  number={9},
  pages={850--863},
  year={1993},
  publisher={IEEE}
}

@article{taha2015efficient,
  title={An efficient algorithm for calculating the exact Hausdorff distance},
  author={Taha, Abdel Aziz and Hanbury, Allan},
  journal={IEEE transactions on pattern analysis and machine intelligence},
  volume={37},
  number={11},
  pages={2153--2163},
  year={2015},
  publisher={IEEE}
}

@inproceedings{aspert2002mesh,
  title={Mesh: Measuring errors between surfaces using the hausdorff distance},
  author={Aspert, Nicolas and Santa-Cruz, Diego and Ebrahimi, Touradj},
  booktitle={Proceedings. IEEE international conference on multimedia and expo},
  volume={1},
  pages={705--708},
  year={2002},
  organization={IEEE}
}

@incollection{alt2003computing,
  title={Computing the Hausdorff distance of geometric patterns and shapes},
  author={Alt, Helmut and Bra{\ss}, Peter and Godau, Michael and Knauer, Christian and Wenk, Carola},
  booktitle={Discrete and computational geometry},
  pages={65--76},
  year={2003},
  publisher={Springer}
}