fix(transistors): pickle dictionaries, not lists
The issue of pickling lists is that if all relations are extracted,
each list is 4 items long. However, if later run is done with less than
the total number of relations (e.g., 3), then the indexes no longer
match up to the 4-item long lists.

To remedy this, this pickles dictionaries, keyed by the name of the
relation, so that the proper ordered lists can be recreated.